backend/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
pydantic>=2.10.4
python-dotenv>=1.0.1
groq>=0.15.0
pandas>=2.2.3
numpy>=2.2.1
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
httpx>=0.28.1
python-multipart>=0.0.20
crewai>=1.0.0
crewai-tools>=0.20.0
twilio>=9.0.0
matplotlib>=3.9.0
